Generalize Connection Manager subscriptions
Generalize the Connection Manager subscriptions to take a Sender that
implements From<ConnectionmanagerNotification>.

This allows for a reduction in threads required in the peer manager and
other users that take a notification and pass it directly to an
alternative sender.

Previously, a thread was needed to read the notifications from the
NotificationIter, only to take that notification and pass it on to
another sender to the PeerManager's background thread.  With this
change, the notification is immediately placed on the sender to the
background thread.

This generalization restructures the subscribers into a SubscriberMap,
where each subscriber has an SubscriberId.  This id can be used to
unsubscribe from notifications if desired.

The original subscribe method has been renamed to subscription_iter.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>